pyproject.toml
[tool.poetry]
name = "DDDBox"
version = "0.1.1"
description = ""
authors = ["Paweł Kucmus <pkucmus@gmail.com>"]
include = ["configs/**/*"]


[tool.poetry.dependencies]
python = "^3.8"
Pillow = "^7.0.0"
numpy = "^1.18.1"
pyyaml = "^5.3"
pyserial = "^3.4"
click = "^7.1.1"
pydantic = "^1.4"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
black = {version = "^19.10b0", allow-prereleases = true}
isort = "^4.3.21"
flake8 = "^3.7.9"
ipython = "^7.13.0"
ipdb = "^0.13.2"
fabric = "^2.5.0"

[tool.poetry.scripts]
dddbox = 'dddbox.main:cli'

[tool.black]
line-length = 79
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
line_length = 79
known_third_party = 'yaml, serial, pillow'
indent = '    '
multi_line_output = 3
default_section = 'FIRSTPARTY'
combine_as_imports = true
use_parentheses = true
force_grid_wrap = 0
include_trailing_comma = true

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"


